requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
colormath==3.0.0
conda==22.11.1
conda-content-trust==0.1.3
conda-libmamba-solver==22.8.1
conda-package-handling==2.0.2
conda_package_streaming==0.8.0
contourpy==1.1.0
cryptography==40.0.2
cycler==0.11.0
datasets==3.1.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.9.0
fonttools==4.42.0
frozenlist==1.5.0
fsspec==2024.9.0
future==0.18.3
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==6.0.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.2
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.12.0
kiwisolver==1.4.4
libmambapy==1.4.2
lzstring==1.0.4
mamba==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclient==0.10.1
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.1.2
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
pluggy==1.0.0
pooch==1.7.0
prompt_toolkit==3.0.50
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.4
pycparser==2.21
pydantic_core==2.27.2
Pygments==2.16.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2025.2
PyYAML==6.0.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.5.2
rich-click==1.6.1
rpds-py==0.20.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scipy==1.10.1
setuptools==67.7.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spectra==0.0.11
stack-data==0.6.3
sympy==1.12
tensorboard==2.12.3
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tinycss2==1.2.1
toolz==0.12.0
torch==2.3.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==2.3.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.0.3
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.40.0
xxhash==3.5.0
yarg==0.1.9
yarl==1.15.2
zipp==3.15.0
zstandard==0.19.0
